requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
chromadb>=0.5.0
openai>=1.50.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.8.0
